Bugfix: Don't show two submenus at same time
On a screen with a mouse pointer (i.e. not a touch screen) if one clicked on a
sumbenu, that submenu would stay open even as one hovered over other submenus.
This completely separates the logic of handling touch screens and screens with
mouse cursors so that they do not interfere with each other.